Finished CellFactory continue to work on sphere.hpp
