+++ 3d/C++/includes/CellFactory.hpp
@@ -5,9 +5,20 @@
 #ifndef CELLUNIVERSE_CELLFACTORY_HPP
 #define CELLUNIVERSE_CELLFACTORY_HPP
 
+#include <string>
+#include <vector>
+#include <map>
+#include <fstream>
+#include <sstream>
+#include "ConfigTypes.hpp"
+#include "Cell.hpp"
+#include "Sphere.hpp"
+#include "types.hpp"
 
 class CellFactory {
-
+public:
+    explicit CellFactory(const BaseConfig& config);
+    CellMap create_cells(const Path &init_params_path, int z_offset = 0, float z_scaling = 1.0);
 };
