+++ 3d/C++/includes/ConfigTypes.hpp
@@ -2,8 +2,8 @@
 #ifndef CONFIGTYPES_HPP
 #define CONFIGTYPES_HPP
 #include <vector>
+#include <random>
 #include <string>
-#include "Sphere.hpp"
 #include "yaml-cpp/yaml.h"
 
 class SimulationConfig {
@@ -48,15 +48,77 @@ public:
     }
 };
 
+class CellParams {
+    //The CellParams class stores the parameters of a particular cell.
+public:
+    std::string name;
+};
+
+class PerturbParams {
+    //Used with a CellConfig to add perturb parameters.
+public:
+    float prob;
+    float mu;
+    float sigma;
+    void explodeParams(const YAML::Node& node) {
+        prob = node["prob"].as<float>();
+        mu = node["mu"].as<float>();
+        sigma = node["sigma"].as<float>();
+    }
+    [[nodiscard]] float get_perturb_offset() const {
+        std::random_device rd; // Obtain a random number from hardware
+        std::mt19937 gen(rd()); // Seed the generator
+        std::uniform_real_distribution<> dis(0.0, 1.0); // Distribution for probability
+
+        if (dis(gen) < prob) {
+            std::normal_distribution<> d(mu, sigma); // Distribution for the Gaussian
+            return d(gen);
+        } else {
+            return mu;
+        }
+    }
+};
+
+class CellConfig {
+    // A pure abstract base class for SphereConfig and BacilliConfig
+public:
+    // pure virtual function for exploding the configuration
+    virtual void explodeConfig(const YAML::Node& node) = 0;
+    virtual ~CellConfig() = 0;
+//    virtual CellConfig& operator=(const CellConfig& other) = 0;
+};
+
+class SphereConfig: public CellConfig {
+public:
+    PerturbParams x{};
+    PerturbParams y{};
+    PerturbParams z{};
+    PerturbParams radius{};
+    double minRadius{};
+    double maxRadius{};
+
+    void explodeConfig(const YAML::Node& node) override
+    {
+        x.explodeParams(node["x"]);
+        y.explodeParams(node["y"]);
+        z.explodeParams(node["z"]);
+        radius.explodeParams(node["radius"]);
+        minRadius = node["minRadius"].as<double>();
+        maxRadius = node["maxRadius"].as<double>();
+    }
+};
 
 class BaseConfig {
 public:
     std::string cellType;
-    // TODO: Change to a template to support more different types of config
-    CellConfig* cell;
+    // TODO: Change to a template to support more different types of config or use abstract class
+    SphereConfig* cell;
     SimulationConfig simulation;
     ProbabilityConfig prob;
     BaseConfig() :cell(nullptr) {};
+    ~BaseConfig() {
+        delete cell;
+    }
     // load the BaseConfig with a YAML node
     void explodeConfig(const YAML::Node& node) {
         cellType = node["cellType"].as<std::string>();
