+++ 3d/C++/includes/types.hpp
@@ -11,6 +11,8 @@
 #include <vector>
 #include <unordered_map>
 #include <filesystem>
+#include <map>
+#include "Sphere.hpp"
 
 enum argKeywords {
     ff = 1,
@@ -28,6 +30,7 @@ typedef std::pair<double, std::function<void(bool)>> CostCallbackPair;
 typedef std::function<void(bool)> CallBackFunc;
 typedef std::unordered_map<std::string, ImageStack> ParamImageMap;
 typedef std::unordered_map<std::string, float> ParamValMap;
+typedef std::map<Path, std::vector<Sphere>> CellMap;
 typedef double Cost;
 namespace fs = std::filesystem;
 typedef std::vector<fs::path> PathVec;
